Add StyleKeeper and StyleSheet components
StyleKeeper is a stateful instance on the Radium root that keeps track
of snippets of CSS. StyleSheet, rendered by said root, renders that CSS
into a style tag. Since the StyleSheet component comes after the
content, calls to populate it in the content’s render methods will
cause it to update before it is rendered on the server.