[Prefix] Add an option to prefix for all user agent
My use case for this, is to prefix for all user agent on the server side
and only for the browser agent on the client side. As pointed out here https://github.com/facebook/react/issues/1979
I think that it's a very good trade-off.
This allow me to remove the user agent from the memoization I use on the `renderToString` method.

The test fails because of https://github.com/rofrischmann/inline-style-prefixer/issues/58.
I'm gonna address it.

Fix https://github.com/FormidableLabs/radium/issues/546.