Move dependencies
I have not tested if this change works, but just wanted to bring it to attention. Services like www.webpackbin.com, and other services, analyzes package.json files and by having a lot of devDependencies in "dependencies" field it potentially breaks 3rd party stuff.

Sorry if these really are dependencies, but can not imagine that is the case? :-)