deps: update karma-webpack
The current version of karma-webpack has a peer dependency
conflict with the local version of webpack. This causes
installs of this repo to fail with npm 7.

Updating karma-webpack to ^3.0.0 resolves this conflict and
all of the tests seem to still pass 🎉

There is a 4.x release of karma-webpack but I figured minimal
improvement lowered the chance of breaking stuff